requirements.txt
Adafruit-Blinka==3.4.1
adafruit-circuitpython-neopixel==4.0.0
Adafruit-PlatformDetect==1.4.4
Adafruit-PureIO==1.0.4
Click==7.0
Flask==1.1.1
Flask-Cors
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pyftdi==0.42.2
pyserial==3.4
pyusb==1.0.2
rpi-ws281x==4.2.3
RPi.GPIO==0.7.0
spidev==3.4
sysv-ipc==1.0.1
Werkzeug==0.16.0
